requirements.txt
fastapi
uvicorn
pydantic
qrcode
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
